Fix syntax of List-ID header (RFC 2919)
This patch fixes the List-ID header to conform to the RFC 2919 [1]
syntax specification. In particular, the '@` is not supported by
the list-id token as is seen in the RFC:

list-id-header = "List-ID:" [phrase] "<" list-id ">" CRLF
list-id = list-label "." list-id-namespace
list-label = dot-atom-text
(from RFC 2822 [2]) dot-atom-text = 1*atext *("." 1*atext)

where atext does not include the '@' character.