Encode e-mails to base64 for Celery
This patch allows 8bit messages to be forwarded as-is by lists.sr.ht.

We've been using str(mail) to convert a parsed message to a string.
However Python will convert any 8bit Content-Transfer-Encoding to
quoted-printable or base64 when formatting the string.

To prevent this, directly use the bytes coming from envelope.content.
However, Celery uses JSON to dispatch tasks data, and Python's stdlib
can't encode bytes.

To workaround this limitation, convert to base64.