API: Rig up (some) legacy webhooks
Still need to rig up patchset and subscription events, but this lays
down the essential riggings and delivers list:create, list:update, and
list:delete.

This uncovered a flaw in core-go's legacy webhooks code which was
addressed by core-go:468752564125e79b5efdfff091d8886dc23e373a and
required the included databse migration. Deleting a list would cascade
to delete any subscriptions associated with that list, which would
prevent list:delete events from being delivered. This sets the list_id
column to null instead of deleting the subscription row, and tweaks the
ordering of tasks in webhook processing to ensure we have a valid list
of subscriptions prior to committing the transaction which deletes the
associated list.