api: fix crash in triggerUserEmailWebhooks
A user can have post permissions for a private list that they cannot
browse. If this happens, the triggerUserEmailWebhooks mutation, which
gets triggered for every incoming email, currently crashes. This, in
turn causes an exception in lists.sr.ht-process in `dispatch_message`,
causing such emails to be never forwarded to list members.

The `emailsByIdLoader` rightfully refuses to load the email, so handle
this by allowing the mutation to return `nil` and return early. It is
arguably more correct to deny user webhooks for emails that are
inaccessible to the user anyways.

The internal mutation is currently only used in one place that does not
use the result, so no further changes are necessary.