Export & Import patch actual statuses
Instead of parsing the latest email having a
X-Sourcehut-Patchset-Update header to determine what is the status
of the patchset; this patch changes the behavior so that the
status of the patch is exported and stored in the generated mbox,
then read on import.

This ensures that changing the status of a patchset in the srht UI,
then exporting and importing, keeps the status.