lists.sr.ht-process: unsubscribe by email, too
In some circumstances, when a user subscribed to a list before creating
an account for the same email address, it can be impossible for the user
to unsubscribe from a list, because the unsubscribe code path only
checks for subscriptions for the user's ID, not the user's email.

This commit makes the unsubscribe code path look for subscriptions for
the user's email as well, if none were found for the ID. While this does
not fix the root cause (doing so is a bit more involved), it does allow
users who are in this situation to successfully unsubscribe.