Let list admins configure rejection of HTML parts
Currently, the rejection of mails that contain a HTML part is coded into
the config. There was a previous attempt to move this into the list
settings, but it was not quite complete.

This commit puts list admins into full control over whether they want
their lists to accept mails with HTML parts, with one exception: we will
always require that a text/plain part is _also_ present, because
otherwise the UI would not know what to do with the mail.

The following is covered:

* Make sure the current value from the config is added to every lists'
  rejected_mimetypes, unless already present - for the main sr.ht
  instance, this is 'text/html'
* Stop using the value from the config in listssrht-lmtp (and drop it
  from the example config)
* Fix the acceptance logic in listssrht-lmtp