api: add mailing list userACL endpoint
In preparation for replacing lmtp agent direct database access with
graphql queries to manage subscriptions in the lmtp agent, we need more
endpoints in the API.

Add endpoint to get the actual mailing list permissions for a given
email address.

This endpoint is only meant to be called internally from the ingress
daemon. Decorate it with @internal.

Signed-off-by: Robin Jarry <robin@jarry.cc>