api: add internal endpoint to get list subscriptions
In preparation for replace lmtp agent direct database access with
graphql queries to manage subscriptions in the lmtp agent, we need more
endpoints in the API.

Update the MailingListSubscription type to have a new subscriber field
(type Entity).

Decorate these new fields as @internal since they should not be exposed
to the public API.

Update the MailingListSubscription go type to reflect these changes.
Notably, change UserID to a pointer to int as the database column is
nullable.

Signed-off-by: Robin Jarry <robin@jarry.cc>