Reimplement the ingress daemon using go
Rename the binary from listssrht-lmtp to ingress/ingress as requested by
Drew since it can be used either with smtp or lmtp. sr.ht-pkgbuilds will
need to be updated accordingly.

No more celery stuff. dowork.Queue is used to avoid blocking the SMTP
ingress (see config.example.ini).

The SMTP server is only processing the strict minimum before passing the
message to one of the ingress workers which takes care of validating the
message, archiving it, looking up the destinations and finally passing
it to another worker via the default core-go email egress queue.

The ingress and egress queue sizes can be configured with:

  [mail]
  egress-queue-size = 512

And:

  [lists.sr.ht::worker]
  ingress-queue-size = 512

The number of workers spawned for both ingress and egress queues can be
configured with:

  [lists.sr.ht::worker]
  queue-workers = 1

If any of the ingress or egress tasks fail, retry executing them 10
times with exponential backoff.

The database is not accessed anymore. The graphql api is used instead.

Most of the work is done in the api daemon behind the archiveMessage
endpoint. The ingress daemon only takes care of the email validation and
dispatch.

Signed-off-by: Robin Jarry <robin@jarry.cc>