ingress: fix corrupted messages after forwarding
For some unknown reason, formatting an email after parsing it with
go-message causes issues. Random line breaks are inserted into the body
of the message. I haven't had the time to properly diagnose what is the
root cause.

To avoid that problem, skip the parsing/reformatting of messages before
forwarding them to subscribers. Simply prepend mailing list headers to
the original buffer. Move "reserved" list headers to the "prohibited"
headers to be handled during the validation phase.

Fixes: d162fddce831 ("Reimplement the ingress daemon using go")
Signed-off-by: Robin Jarry <robin@jarry.cc>