archive: fix threading of emails received out of order
When receiving a patch series out of order and the cover letter is
received last, all existing messages that are children of the cover
letter are not updated to have their parent_id and thread_id updated to
the id of the cover letter message.

This breaks patchset creation and threading in the web UI.

The issue occurs because existing children are searched with the
received Message-ID without stripping it from the <> brackets. For
historical reasons, the Message-ID is stored as is in the database (with
the angle brackets) and In-Reply-To is stored after stripping the angle
brackets.

Adjust the query for children by stripping the angle brackets from the
received Message-ID.

Fixes: cde342868e5d ("api: Implement mailing list spool imports")
Signed-off-by: Robin Jarry <robin@jarry.cc>