listssrht: don't let the existence of random users poison from: searches
Right now, the "from:foo" search filter checks if a "foo" SourceHut user
exists, and if so, looks for emails from that user's email address; if
not, it looks for emails whose from contain "foo".

The problem reported in ticket 207 is that the natural way of searching
for mails sent by Tim, from:tim, will only do what we want as long as
there's no ~tim user defined in the system.

This patch only resolves the foo in "from:foo" if it starts with a ~;
otherwise, it goes straight to fuzzy matching. I realize that it's a
behaviour change, but it makes the semantics of "from:" search only
depend on what's in the list, not things end users are oblivious to.

Fixes: https://todo.sr.ht/~sircmpwn/lists.sr.ht/207
Signed-off-by: Simon Martin <simon@nasilyan.com>