archive: keep track of last successful message archival in a list
The "Last active:" information in https://lists.sr.ht is wrong IMHO: it
uses the list's "updated" attribute, that is only updated when the list
characteristics are changed, not when new messages are processed.

To fix this, this patch adds a new column to the list table to store the
time of the last successful archival operation on that list, shows it in
"Last active:", and uses it for sorting in the pages that show lists.

Signed-off-by: Simon Martin <simon@nasilyan.com>