api,listssrht: allow patch submitters to update their own patches
Unless you're the list owner or have moderation rights, it's currently
impossible to update the status of your own patches.

This patch allows patch submitters to update the status of **their own**
patches (transitions to Approved and Applied are still only possible for
moderators) via the UI (if they have a SourceHut account) and via email
(whether they have a SourceHut account or not).

Fixes: https://todo.sr.ht/~sircmpwn/lists.sr.ht/193
Signed-off-by: Simon Martin <simon@nasilyan.com>