api: don't use int32 to represent database ids
I noticed some uses of pq.Int32Array, that sounds like an incident
waiting to happen in a few (many?) years.

This patch changes those uses to pq.Array, and as a consequence a few
int32s to int in various callers.

What do people think, and what kind of testing (beyond checking that we
continue to process inbound messages and "large" reply chains - which I
did with sr.ht-container-compose) would you advise for such a change?

Signed-off-by: Simon Martin <simon@nasilyan.com>