listssrht: report correct author for patches submitted by someone else
When a patch is submitted on behalf of its author, the mail generated by
git send-email will have the submitter information in the From header,
and the author information in the first line of the body.

lists.sr.ht always uses the From header to determine the author, and
will be wrong in such a case; this patch fixes this.

Fixes: https://todo.sr.ht/~sircmpwn/lists.sr.ht/202
Signed-off-by: Simon Martin <simon@nasilyan.com>