ingress: fix reply-only permission logic
Currently, configuring a list with reply-only permissions is not
possible. Replies get rejected even though they have a correct
In-Reply-To header pointing to an existing message on the list.

This happens because we wrongfully check if the current message is
already in the database (effectively checking for duplicate messages).
This duplicate Message-ID check is already performed by the API backend.

We must check if the message identified by the In-Reply-To header is
already present in the archive to determine if the ingress message is
a reply or not.

Fixes: 87367d14c049 ("Reapply "Reimplement the ingress daemon using go"")
Reported-by: Maximilian Moehl <maximilian@moehl.eu>
Signed-off-by: Robin Jarry <robin@jarry.cc>