list: fix thread orphans due to parent_id computation race condition
This ticket shows that if patches in a series and its cover letter are
processed in parallel, it's possible that the patch's parent_id is not
properly computed due to a race condition. The consequence is that
(1) the messages are not properly threaded, and (2) the patch is not
considered complete, and not shown in the "Patches" view.

This patch fixes this issue by using an advisory lock to make sure that
on a given list, at most one message is being archived at any time.

Fixes: https://todo.sr.ht/~sircmpwn/lists.sr.ht/215
Suggested-by: Conrad Hoffmann <ch@bitfehler.net>
Signed-off-by: Simon Martin <simon@nasilyan.com>