listssrht: properly handle angle brackets when searching by header
Using angle brackets when searching by Message-Id or In-Reply-To gives
no result, even when there are matching messages. That's because we
always search in the (JSON encoded) headers column, even for headers
that are also stored in dedicated columns.

This patch fixes this by using the dedicated column when applicable, and
adapting to our mixed management of angle brackets.

Signed-off-by: Simon Martin <simon@nasilyan.com>