Re-align DB schema with SQLAlchemy behavior
Currently, if one tries to delete an email that is the cover letter of a
patchset, SQLAlchemy will complain about a circular dependency:

    sqlalchemy.exc.CircularDependencyError: Circular dependency
    detected. (DeleteState(<Email at 0x7f330b4bcf80>),
    DeleteState(<Patchset at 0x7f330b4bd520>))

The investigation into this lead to the discovery that there is a
significant divergence between the DB schema and how SQLAlchemy behaves.
In the schema, `email.patchset_id` is currently set to `ON DELETE
CASCADE`, which expresses that if a _patchset_ is deleted, _all_ emails
that reference it should be deleted as well.

This a) seems wrong and b) is not what actually happens. I verified that
a simple `DELETE FROM patchset` SQL query indeed causes all emails
referencing the patchset to be deleted. However, all email and patchset
deletion currently still happens through SQLAlchemy, so it obviously
does some magic before performing the actual deletion that prevents this
cascade from happening.

This commit tries to:

1. Re-align the schema with the current SQLAlchemy behavior, meaning the
   deletion of a patchset does _not_ trigger deletion of all emails
   referencing it - this will become important once deletion via GQL
   gets implemented
2. Have SQLAlchemy stop doing magic, so that cascades are handled by the
   database, not by SQLAlchemy (and hence potentially differently from
   what the schema says)

I can only half-explain why SQLAlchemy handled deletion so differently
from how one would expect given the database schema, but I think it
boils down to this paragraph from [1] (emphasis mine):

> In order to use ON DELETE foreign key cascades in conjunction with
> relationship(), it’s important to note first and foremost that the
> relationship.cascade setting *must still be configured* to match the
> desired "delete" or "set null" behavior (using delete cascade or
> leaving it omitted)...

The next paragraph states:

> There is then an additional option on relationship() which indicates
> the degree to which the ORM should try to run DELETE/UPDATE operations
> on related rows itself, vs. how much it should rely upon expecting the
> database-side FOREIGN KEY constraint cascade to handle the task; this
> is the relationship.passive_deletes parameter and it accepts options
> False (the default), True and "all".

So in order to avoid such a schema/ORM divergence, and also to fix the
circular dependency issue (which is a pure ORM problem, not a SQL
problem), I added the `passive_deletes=True` [2] to the relevant
relationship, to tell SQLAlchemy to let the database sort it out.

With this, I can say that schema and ORM behavior seem to be re-aligned,
and the exception that lead to all this no longer occurs. I am _not_
100% sure, even after reading the documentation, whether there should
_still_ be a `cascade` parameter to the relationship. But I would prefer
spending time removing this code over spending even more time trying to
understand SQLAlchemy :) (did I mention it seems to work?)

[1]: https://docs.sqlalchemy.org/en/20/orm/cascades.html#using-foreign-key-on-delete-cascade-with-orm-relationships
[2]: https://docs.sqlalchemy.org/en/20/orm/relationship_api.html#sqlalchemy.orm.relationship.params.passive_deletes