Rename "envelope" to "raw message"
In the context of emails, the envelope is usually the part of the SMTP
transaction that specifies sender and receiver, which infamously is
completely decoupled from the contents of the actual email. Hence, the
name envelope to refer to the raw message is a bit confusing.

This commit changes the nomenclature from "envelope" to "raw message" in
all relevant contexts.