server: allow mbox to be empty for empty mailing list
I have empty lists in my account, and noticed that `hut export` shows an
error for some of them, while they're perfectly valid.

This patch allows mbox to be empty for lists, but not for threads or
patchsets (they should not have an ID in the first place if they don't
exist, so we can't try to export them).

Signed-off-by: Simon Martin <simon@nasilyan.com>