patchset: link to superseding patch, if any
We now keep track of the superseded/superseding relationship when the
patch superseding is automatically done, but don't expose it anywhere.

This adds an info block to the patchset view for automatically
superseded patchsets to give users a link to the superseding one.

References: https://todo.sr.ht/~sircmpwn/lists.sr.ht/102
Signed-off-by: Simon Martin <simon@nasilyan.com>